# SWAGGER_OPENAPI/datavault-broker-openapi.yaml
openapi: 3.1.0
info:
  title: DataVault Broker API
  description: |-
    API documentation for the DataVault Broker module.
    All endpoints require authentication via the X-UserID and X-Client-Key headers.
  version: 0.0.0 (4b97dc0)
  x-generated-at: 2026-03-12T10:56:37.585546
servers:
- url: http://localhost
  description: Generated server url
security:
- X-UserID-Header: []
- X-Client-Key-Header: []
...
paths:
...
  /admin/users:
    put:
      tags:
      - admin-users-controller
      summary: Edit a DataVault User
      description: Updates an existing user in the DataVault system.
      operationId: editUser
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      tags:
      - admin-users-controller
      summary: Create a new DataVault User
      description: Creates a new user in the DataVault system.
      operationId: addUser
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/retentionpolicies:
    put:
      tags:
      - admin-retention-policies-controller
      summary: Edit a Retention Policy
      description: Updates an existing Retention Policy.
      operationId: editRetentionPolicy
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      - name: X-Client-Key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRetentionPolicy"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRetentionPolicy"
    post:
      tags:
      - admin-retention-policies-controller
      summary: Create a new Retention Policy
      description: Creates a new Retention Policy in the DataVault system.
      operationId: addRetentionPolicy
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      - name: X-Client-Key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRetentionPolicy"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRetentionPolicy"
...
  /users:
    get:
      tags:
      - users-controller
      summary: Gets a list of all Users
      description: Retrieves a list of all users in the DataVault system.
      operationId: getUsers
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      tags:
      - users-controller
      summary: Create a new DataVault User
      description: Creates a new user in the DataVault system.
      operationId: addUser_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /users/{userId}:
    get:
      tags:
      - users-controller
      summary: Get a specific DataVault User
      description: Retrieves details for a specific user by their ID.
      operationId: getUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /admin/users/search:
    get:
      tags:
      - admin-users-controller
      summary: Search Users
      description: Searches for users based on a query string.
      operationId: getUsers_1
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      - name: query
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        password:
          type: string
        email:
          type: string
        properties:
          type: object
          additionalProperties:
            type: string
...
    CreateRetentionPolicy:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: Uh... the id
        name:
          type: string
          description: A name for the new Retention Policy
        description:
          type: string
          description: A description of Retention Policy
        engine:
          type: string
          description: Engine for Retention Policy (deprecated
        sort:
          type: string
          description: Sorting order for Retention Policy (deprecated)
        url:
          type: string
          description: URL for Retention Policy
        minRetentionPeriod:
          type: integer
          format: int32
          description: Minimum retention period in years
        extendUponRetrieval:
          type: boolean
          description: Extend the expiry date of a vault if a deposit is retrieved
        minDataRetentionPeriod:
          type: string
          description: Minimum Date Retention Period (deprecated)
        inEffectDate:
          type: string
          format: date
          description: In Effect Date
        endDate:
          type: string
          format: date
          description: End Date
        dataGuidanceReviewed:
          type: string
          format: date
          description: Date Guidance Reviewed
        action:
          type: string
          description: Form Action
        policyInfo:
          type: string
...
  securitySchemes:
    X-UserID-Header:
      type: apiKey
      description: "User ID header, corresponds to `org.datavaultplatform.common.util.Constants.HEADER_USER_ID`."
      name: X-UserID
      in: header
    X-Client-Key-Header:
      type: apiKey
      description: "Client Key header, corresponds to `org.datavaultplatform.common.util.Constants.HEADER_CLIENT_KEY`."
      name: X-Client-Key
      in: header